refactor: eliminate code duplication across simulator tools (#3, #5)
Extract duplicated code to shared utilities, removing 144 lines of
identical code across simulator tools. This completes Batch 2 of the
code quality improvements epic.

## Part 1: Extract Schema Definitions (Todo 001)
Create shared schema module for common simulator options:

NEW FILE: src/mcp/tools/simulator/shared-schemas.ts
- simulatorCommonOptions: 9 common fields (platform, scheme, simulatorId, etc.)
- projectWorkspaceOptions: Mutually exclusive project/workspace paths
- Comprehensive JSDoc documentation

MODIFIED: build_sim.ts, build_run_sim.ts
- Import and use shared schemas via spread operator
- Remove 54 lines of duplicated baseOptions per file
- Net reduction: 108 LOC of duplication

## Part 2: Extract Platform Mapping (Todo 002)
Create platform mapping utility function:

NEW FILE: src/utils/platform-utils.ts
- mapPlatformStringToEnum(): Maps platform strings to XcodePlatform enum
- Handles defaults and unknown platforms gracefully
- JSDoc with usage examples

MODIFIED: test_sim.ts, build_sim.ts, build_run_sim.ts
- Import and use mapPlatformStringToEnum utility
- Remove 6 lines of duplicated platformMap per file
- Net reduction: 18 LOC of duplication

## Part 3: Extract Warning Logic (Todo 003)
Create simulator validation utility:

NEW FILE: src/utils/simulator-validation.ts
- logUseLatestOSWarning(): Warns when useLatestOS used with simulatorId
- Comprehensive JSDoc explaining why warning exists

MODIFIED: test_sim.ts, build_sim.ts, build_run_sim.ts
- Import and use logUseLatestOSWarning utility
- Remove 6 lines of duplicated warning logic per file
- Net reduction: 18 LOC of duplication

## Build System Fix
MODIFIED: build-plugins/plugin-discovery.js
- Exclude shared-schemas.ts from tool discovery
- Prevents treating utility files as MCP tools
- Fixes "missing default export" error

## Impact Summary
- Removed: 144 lines of code duplication
- Added: 3 new utility modules with proper documentation
- Improved: Maintainability (single source of truth)
- Improved: Build system (correct tool discovery)

## Quality Status
✅ Typecheck: Zero errors
✅ Lint: Zero errors (1 pre-existing warning unrelated)
✅ Build: Successful (13 workflows, 85 tools)
✅ Tests: 1192 passing (29 pre-existing failures from Batch 1)

Related: Epic #3, Issue #5
Todos: 001, 002, 003
