+++ src/mcp/tools/simulator/build_sim.ts
@@ -9,60 +9,19 @@
 import { z } from 'zod';
 import { log } from '../../../utils/logging/index.ts';
 import { executeXcodeBuildCommand } from '../../../utils/build/index.ts';
-import { ToolResponse, XcodePlatform } from '../../../types/common.ts';
+import { ToolResponse } from '../../../types/common.ts';
 import type { CommandExecutor } from '../../../utils/execution/index.ts';
 import { getDefaultCommandExecutor } from '../../../utils/execution/index.ts';
 import { createSessionAwareTool } from '../../../utils/typed-tool-factory.ts';
 import { nullifyEmptyStrings } from '../../../utils/schema-helpers.ts';
+import { simulatorCommonOptions, projectWorkspaceOptions } from './shared-schemas.ts';
+import { logUseLatestOSWarning } from '../../../utils/simulator-validation.ts';
+import { mapPlatformStringToEnum } from '../../../utils/platform-utils.ts';
 
 // Unified schema: XOR between projectPath and workspacePath, and XOR between simulatorId and simulatorName
-const baseOptions = {
-  scheme: z.string().describe('The scheme to use (Required)'),
-  platform: z
-    .enum(['iOS Simulator', 'watchOS Simulator', 'tvOS Simulator', 'visionOS Simulator'])
-    .optional()
-    .default('iOS Simulator')
-    .describe('Target simulator platform (defaults to iOS Simulator)'),
-  simulatorId: z
-    .string()
-    .optional()
-    .describe(
-      'UUID of the simulator (from list_sims). Provide EITHER this OR simulatorName, not both',
-    ),
-  simulatorName: z
-    .string()
-    .optional()
-    .describe(
-      "Name of the simulator (e.g., 'iPhone 16'). Provide EITHER this OR simulatorId, not both",
-    ),
-  configuration: z.string().optional().describe('Build configuration (Debug, Release, etc.)'),
-  derivedDataPath: z
-    .string()
-    .optional()
-    .describe('Path where build products and other derived data will go'),
-  extraArgs: z.array(z.string()).optional().describe('Additional xcodebuild arguments'),
-  useLatestOS: z
-    .boolean()
-    .optional()
-    .describe('Whether to use the latest OS version for the named simulator'),
-  preferXcodebuild: z
-    .boolean()
-    .optional()
-    .describe(
-      'If true, prefers xcodebuild over the experimental incremental build system, useful for when incremental build system fails.',
-    ),
-};
-
 const baseSchemaObject = z.object({
-  projectPath: z
-    .string()
-    .optional()
-    .describe('Path to .xcodeproj file. Provide EITHER this OR workspacePath, not both'),
-  workspacePath: z
-    .string()
-    .optional()
-    .describe('Path to .xcworkspace file. Provide EITHER this OR projectPath, not both'),
-  ...baseOptions,
+  ...projectWorkspaceOptions,
+  ...simulatorCommonOptions,
 });
 
 const baseSchema = z.preprocess(nullifyEmptyStrings, baseSchemaObject);
@@ -91,24 +50,11 @@ async function _handleSimulatorBuildLogic(
   const projectType = params.projectPath ? 'project' : 'workspace';
   const filePath = params.projectPath ?? params.workspacePath;
 
-  // Map platform string to XcodePlatform enum
-  const platformMap: Record<string, XcodePlatform> = {
-    'iOS Simulator': XcodePlatform.iOSSimulator,
-    'watchOS Simulator': XcodePlatform.watchOSSimulator,
-    'tvOS Simulator': XcodePlatform.tvOSSimulator,
-    'visionOS Simulator': XcodePlatform.visionOSSimulator,
-  };
-
-  const platform = platformMap[params.platform ?? 'iOS Simulator'] ?? XcodePlatform.iOSSimulator;
+  const platform = mapPlatformStringToEnum(params.platform);
   const platformName = params.platform ?? 'iOS Simulator';
 
   // Log warning if useLatestOS is provided with simulatorId
-  if (params.simulatorId && params.useLatestOS !== undefined) {
-    log(
-      'warning',
-      `useLatestOS parameter is ignored when using simulatorId (UUID implies exact device/OS)`,
-    );
-  }
+  logUseLatestOSWarning(params.simulatorId, params.useLatestOS);
 
   log(
     'info',
